Block Multiple Fainted

+++ src/module/actor/base.js
@@ -851,10 +851,12 @@ class PTUActor extends Actor {
 
         if (data.system?.health?.value !== undefined) {
             if (data.system.health.value <= 0 && game.settings.get("ptu", "automation.autoFaint")) {
+                if(this.primaryUpdater?.id !== game.user.id) return;
                 const fainted = this.conditions.bySlug("fainted");
                 if (fainted.length === 0) PTUCondition.FromEffects([{ id: "fainted" }]).then(items => this.createEmbeddedDocuments("Item", items));
             }
             else if (data.system.health.value > 0 && game.settings.get("ptu", "automation.autoFaintRecovery")) {
+                if(this.primaryUpdater?.id !== game.user.id) return;
                 const fainted = this.conditions.bySlug("fainted");
                 if (fainted.length > 0) fainted.forEach(f => f.delete());
             }
